+++ index.d.ts
@@ -1,21 +1,18 @@
-// Type definitions for alawmulaw 5.0
+// Type definitions for alawmulaw 6.0.0
 // Project: https://github.com/rochars/alawmulaw
-// Definitions by: Rafael S. Rocha <https://github.com/rochars>
+// Definitions by: Rafael da Silva Rocha <https://github.com/rochars>
 // Definitions: https://github.com/rochars/alawmulaw
 
-export = alawmulaw;
+export module alaw {
+    function decode(samples: Uint8Array): Int16Array;
+    function decodeSample(sample: number): number;
+    function encode(samples: Int16Array): Uint8Array;
+    function encodeSample(sample: number): number;
+}
 
-declare const alawmulaw: {
-    alaw: {
-        decode(samples: Uint8Array): Int16Array;
-        decodeSample(sample: number): number;
-        encode(samples: Int16Array): Uint8Array;
-        encodeSample(sample: number): number;
-    };
-    mulaw: {
-        decode(samples: Uint8Array): Int16Array;
-        decodeSample(sample: number): number;
-        encode(samples: Int16Array): Uint8Array;
-        encodeSample(sample: number): number;
-    };
-};
+export module mulaw {
+    function decode(samples: Uint8Array): Int16Array;
+    function decodeSample(sample: number): number;
+    function encode(samples: Int16Array): Uint8Array;
+    function encodeSample(sample: number): number;
+}