+++ test/loader.js
@@ -1,36 +1,20 @@
 /**
- * Copyright (c) 2017-2018 Rafael da Silva Rocha.
+ * Copyright (c) 2017-2019 Rafael da Silva Rocha.
  * https://github.com/rochars/alawmulaw
  *
  */
 
 let alawmulaw;
 
-// Browser
-if (process.argv[3] == '--min') {
-    console.log('browser');
-    require('browser-env')();
-    require('../dist/alawmulaw.min.js');
-    alawmulaw = window.alawmulaw;
-
-// UMD
-} else if (process.argv[3] == '--umd') {
+if (process.argv[3] == '--umd') {
 	console.log('umd');
-	alawmulaw = require('../dist/alawmulaw.umd.js');
-
-// CommonJS
-} else if (process.argv[3] == '--cjs') {
-	console.log('cjs');
-	alawmulaw = require('../dist/alawmulaw.cjs.js');
-
-// esm
-} else if (process.argv[3] == '--esm') {
-	console.log('esm');
 	alawmulaw = require('../dist/alawmulaw.js');
 
 // source
 } else {
 	console.log('source');
+	require = require("esm")(module);
+	global.module = module;
 	alawmulaw = require('../index.js');
 }